fix: comparing string to array of integers
This seems to be something that worked initially but because of a change
in the default comparison on sorted-array-functions library, it surfaced
the incorrect comparison of an array of integers and a string. The life
changing commit is here:
https://github.com/mafintosh/sorted-array-functions/commit/07672c2b272366cb2e873ef4474298d3c192e792.

The `master` branch works with sorted-array-functions@1.2.0 (pinned) but
fails at 1.3.0.

I chose not to save the result of the `parseInt` for two reasons:
  * a little bit into the code, every address in the stack is run through
    `parseInt` so trying to avoid wierdness
  * the string result is ultimately sent to the callback function